added DEM caching

+++ decision_tree/api_utils.py
@@ -1,3 +1,5 @@
+import shutil
+
 import requests
 import yaml
 import pandas as pd
@@ -26,7 +28,7 @@ from tm_api_utils import pull_tm_api_data
 
 from decision_tree.constants import TM_STAGING_URI, TM_PROD_URI
 from decision_tree.s3_utils import upload_to_s3
-from decision_tree.tools import get_project_root, convert_to_os_path
+from decision_tree.tools import get_project_root, convert_to_os_path, file_exists_and_fresh, create_folder_if_not_exists
 
 
 def get_ids(params):
@@ -140,28 +142,13 @@ def get_tm_feats(params, secrets, geojson_dir, tm_outfile, project_ids):
 
     if tm_outfile:
         dir = os.path.dirname(tm_outfile)
-        _create_folder_if_not_exists(dir)
+        create_folder_if_not_exists(dir)
         outfile_path = os.path.join(get_project_root(), tm_outfile)
         with open(outfile_path, "w") as f:
             json.dump(all_results, f, indent=4)
         print(f"Results saved to {outfile_path}")
 
     return all_results
-
-
-def _create_folder_if_not_exists(folder_path):
-    """
-    Creates a folder if it does not already exist.
-
-    Args:
-        folder_path (str): The path of the folder to create.
-    """
-    try:
-        # os.makedirs with exist_ok=True will not raise an error if the folder exists
-        os.makedirs(folder_path, exist_ok=True)
-        print(f"Folder ready at: {folder_path}")
-    except OSError as e:
-        print(f"Error creating folder '{folder_path}': {e}")
 
 
 def calculate_high_slope_area(slope_raster, polygon, threshold=20):
@@ -247,14 +234,27 @@ def opentopo_pull_wrapper(params, secrets, geojson_dir, feats_df, process_in_utm
             slope_path = slope_tmp.name
 
         try:
-            # Download the DEM
-            response = requests.get(dem_url, stream=True, params=query)
-            if response.status_code != 200:
-                raise Exception(f"Error downloading DEM for project {name}: {response.text}")
+            cached_dem_dir = convert_to_os_path(project_data_dir, "dem_cache")
+            create_folder_if_not_exists(cached_dem_dir)
 
-            with open(dem_path, 'wb') as f:
-                for chunk in response.iter_content(8192):  # 8k generally works well for chunking of data.
-                    f.write(chunk)
+            cached_dem_file_path = os.path.join(cached_dem_dir, f"{name}_{data_version}.tif")
+            fresh_dem_exists = file_exists_and_fresh(cached_dem_file_path, max_age_days=1)
+
+            if fresh_dem_exists:
+                # read cached file
+                shutil.copyfile(cached_dem_file_path, dem_path)
+            else:
+                # Download the DEM
+                response = requests.get(dem_url, stream=True, params=query)
+                if response.status_code != 200:
+                    raise Exception(f"Error downloading DEM for project {name}: {response.text}")
+
+                with open(dem_path, 'wb') as f:
+                    for chunk in response.iter_content(8192):  # 8k generally works well for chunking of data.
+                        f.write(chunk)
+
+                # write to cache file
+                shutil.copyfile(dem_path, cached_dem_file_path)
 
             if process_in_utm_coordinates:
                 site_polygons, slope_raster = _prepare_utm_features(project_polygons, dem_path)
@@ -298,7 +298,7 @@ def opentopo_pull_wrapper(params, secrets, geojson_dir, feats_df, process_in_utm
             project_results_df = pd.concat(this_project, ignore_index=True)
 
             stats_dir = os.path.dirname(stat_path)
-            _create_folder_if_not_exists(stats_dir)
+            create_folder_if_not_exists(stats_dir)
             project_results_df.to_csv(stat_path, index=False)
 
             slope_raster.close()
